wip: abstract out updating buy quote
